initial burndown to get things loading again
