+++ test/fixtures/clinics.yml
@@ -1,34 +1,34 @@
 # A set of fake clinics with cost, geo, and naf data to use as stubs
-planned_parenthood_oakland:
-  _id: abgd1
-  street_address: 1001 Broadway
-  city: Oakland
-  state: CA
-  zip: 94607
-  accepts_naf: false
-  accepts_medicaid: false
-  gestational_limit: 139
-  costs_9wks: 425
-  costs_12wks: 475
-  costs_18wks: 975
-  costs_24wks: null
-  costs_30wks: null
+# planned_parenthood_oakland:
+#   _id: abgd1
+#   street_address: 1001 Broadway
+#   city: Oakland
+#   state: CA
+#   zip: 94607
+#   accepts_naf: false
+#   accepts_medicaid: false
+#   gestational_limit: 139
+#   costs_9wks: 425
+#   costs_12wks: 475
+#   costs_18wks: 975
+#   costs_24wks: null
+#   costs_30wks: null
 
-planned_parenthood_san_fran:
-  _id: abgd2
-  street_address: 2430 Folsom
-  city: San Francisco
-  state: CA
-  zip: 94110
-  accepts_naf: false
-  accepts_medicaid: false
-  gestational_limit: 111
-  costs_9wks: 300
-  costs_12wks: 325
-  costs_18wks: null
-  costs_18wks: null
-  costs_24wks: null
-  costs_30wks: null
+# planned_parenthood_san_fran:
+#   _id: abgd2
+#   street_address: 2430 Folsom
+#   city: San Francisco
+#   state: CA
+#   zip: 94110
+#   accepts_naf: false
+#   accepts_medicaid: false
+#   gestational_limit: 111
+#   costs_9wks: 300
+#   costs_12wks: 325
+#   costs_18wks: null
+#   costs_18wks: null
+#   costs_24wks: null
+#   costs_30wks: null
 
 castro_family_planning:
   _id: abgd3
