+++ test/test_helper.rb
@@ -83,23 +83,23 @@ class ActionDispatch::IntegrationTest
   # for controllers
 end
 
-# Mirrors, more or less the DCAF class, and some slight rails-y model
-# behavior for attributes
-class Clinic
-  attr_accessor :_id, :name, :street_address, :city,
-                :state, :zip, :accepts_naf,
-                :accepts_medicaid, :gestational_limit,
-                :costs_9wks, :costs_12wks, :costs_18wks,
-                :costs_24wks, :costs_30wks
+# # Mirrors, more or less the DCAF class, and some slight rails-y model
+# # behavior for attributes
+# class Clinic
+#   attr_accessor :_id, :name, :street_address, :city,
+#                 :state, :zip, :accepts_naf,
+#                 :accepts_medicaid, :gestational_limit,
+#                 :costs_9wks, :costs_12wks, :costs_18wks,
+#                 :costs_24wks, :costs_30wks
 
-  def initialize(clinic_hash)
-    clinic_hash.each_pair { |k, v| instance_variable_set("@#{k}".to_sym, v) }
-  end
+#   def initialize(clinic_hash)
+#     clinic_hash.each_pair { |k, v| instance_variable_set("@#{k}".to_sym, v) }
+#   end
 
-  def attributes
-    vars = instance_variables.map do |val|
-      { val.to_s.delete('@').to_sym => instance_variable_get(val) }
-    end
-    vars.reduce(&:merge)
-  end
-end
+#   def attributes
+#     vars = instance_variables.map do |val|
+#       { val.to_s.delete('@').to_sym => instance_variable_get(val) }
+#     end
+#     vars.reduce(&:merge)
+#   end
+# end
